+++ src/dht/mod.rs
@@ -9,3 +9,6 @@ pub use dht_settings::*;
 pub mod dht_event;
 
 mod socket;
+
+/// Functions that use [DHT](crate::dht::DHT) to perform high-level operations on the network.
+pub mod operations;