+++ src/storage/mod.rs
@@ -1,4 +1,4 @@
-mod buckets;
+pub mod buckets;
 pub mod node_bucket_storage;
 pub mod node_wrapper;
 pub mod outbound_request_storage;